// src/shared/sdk/chronos/models/BookingWithDetails.ts
/**
 * Chronos API
 * API documentation for the Chronos application
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Booking } from '../models/Booking';
import { Service } from '../models/Service';
import { HttpFile } from '../http/http';

export class BookingWithDetails {
    'booking': Booking;
    'id': number;
    'magicLink'?: string | null;
    'service': Service;
    'staffId': number;
    'userEmail': string;
    'userName': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "booking",
            "baseName": "booking",
            "type": "Booking",
            "format": ""
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "magicLink",
            "baseName": "magic_link",
            "type": "string",
            "format": ""
        },
        {
            "name": "service",
            "baseName": "service",
            "type": "Service",
            "format": ""
        },
        {
            "name": "staffId",
            "baseName": "staff_id",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "userEmail",
            "baseName": "user_email",
            "type": "string",
            "format": ""
        },
        {
            "name": "userName",
            "baseName": "user_name",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return BookingWithDetails.attributeTypeMap;
    }

    public constructor() {
    }
}