// src/shared/sdk/chronos/models/ServiceAddon.ts
/**
 * Chronos API
 * API documentation for the Chronos application
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ServiceAddon {
    'currency': string;
    'description'?: string | null;
    'id': string;
    'isAvailable': boolean;
    'name': string;
    'price': number;
    'requiresQuote': boolean;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "currency",
            "baseName": "currency",
            "type": "string",
            "format": ""
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string",
            "format": ""
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string",
            "format": ""
        },
        {
            "name": "isAvailable",
            "baseName": "isAvailable",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "price",
            "baseName": "price",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "requiresQuote",
            "baseName": "requiresQuote",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ServiceAddon.attributeTypeMap;
    }

    public constructor() {
    }
}